support for binary classification

+++ src/optimization.py
@@ -71,7 +71,7 @@ def train_stochastic(dataloader, model, optimizer, criterion, epoch, reg=1, norm
 
         y_pred = model(t_x).squeeze()
 
-        loss = criterion(y_pred, t_y) / len(t_x)
+        loss = criterion(y_pred, t_y.float()) / len(t_x)
 
         if reg > 0:
 
@@ -93,7 +93,6 @@ def train_stochastic(dataloader, model, optimizer, criterion, epoch, reg=1, norm
 
         if monitor:
             monitor.write(model, i + last_iter, train={"Loss": loss.detach()})
-
 def evaluate(dataloader, model, criterion, epoch=None, monitor=None):
 
     model.eval()
